split domain to src/header, add __add__ operator overload
